El zoom ya funciona para varios patos a la vez sin afectar al fondo.

+++ client/Items/ClientDuck.h
@@ -19,8 +19,8 @@ public:
     void draw(Renderer& renderer);
     bool checkCollision(Rect rect);
     uint8_t getId() const;
-    float getPosX();
-    float getPosY();
+    int getPosX();
+    int getPosY();
 
 private:
     uint8_t idDuck;
